Move Ai components from RequestPanel to their own files
