+++ frontend/src/pages/RequestorPage/RequestPanel/RequestPanel.tsx
@@ -1,37 +1,10 @@
-import SparkleWand from "@/assets/SparkleWand.svg";
-import { KeyboardShortcutKey } from "@/components/KeyboardShortcut";
 import { Button } from "@/components/ui/button";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuLabel,
-  DropdownMenuRadioGroup,
-  DropdownMenuRadioItem,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
 import { Tabs } from "@/components/ui/tabs";
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip";
 import { useToast } from "@/components/ui/use-toast";
 import { useIsSmScreen } from "@/hooks";
-import { cn, isMac } from "@/utils";
-import {
-  CaretDownIcon,
-  Cross2Icon,
-  EraserIcon,
-  InfoCircledIcon,
-} from "@radix-ui/react-icons";
-import {
-  Dispatch,
-  SetStateAction,
-  useCallback,
-  useEffect,
-  useState,
-} from "react";
+import { cn } from "@/utils";
+import { EraserIcon, InfoCircledIcon } from "@radix-ui/react-icons";
+import { Dispatch, SetStateAction } from "react";
 import { Resizable } from "react-resizable";
 import { CodeMirrorJsonEditor } from "../Editors";
 import { FormDataForm } from "../FormDataForm";
@@ -39,7 +12,7 @@ import { KeyValueForm, KeyValueParameter } from "../KeyValueForm";
 import { PathParamForm } from "../PathParamForm/PathParamForm";
 import { ResizableHandle } from "../Resizable";
 import { CustomTabTrigger, CustomTabsContent, CustomTabsList } from "../Tabs";
-import { AiTestingPersona, FRIENDLY, HOSTILE } from "../ai";
+import { AiTestingPersona } from "../ai";
 import { useResizableWidth, useStyleWidth } from "../hooks";
 import type {
   RequestBodyType,
@@ -47,11 +20,12 @@ import type {
   RequestsPanelTab,
 } from "../reducer";
 import { WebSocketState } from "../useMakeWebsocketRequest";
+import { AiDropDownMenu } from "./AiDropDownMenu";
+import { AIGeneratedInputsBanner } from "./AiGeneratedInputsBanner";
 import {
   RequestBodyTypeDropdown,
   RequestBodyTypeDropdownProps,
 } from "./RequestBodyCombobox";
-
 import "./styles.css";
 
 type RequestPanelProps = {
@@ -407,59 +381,6 @@ export function PanelSectionHeader({
   );
 }
 
-type AIGeneratedInputsBannerProps = {
-  showAiGeneratedInputsBanner: boolean;
-  setShowAiGeneratedInputsBanner: Dispatch<SetStateAction<boolean>>;
-  setIgnoreAiInputsBanner: Dispatch<SetStateAction<boolean>>;
-};
-
-function AIGeneratedInputsBanner({
-  showAiGeneratedInputsBanner,
-  setShowAiGeneratedInputsBanner,
-  setIgnoreAiInputsBanner,
-}: AIGeneratedInputsBannerProps) {
-  const onClose = () => {
-    setShowAiGeneratedInputsBanner(false);
-  };
-
-  const onDontShowAgain = () => {
-    setIgnoreAiInputsBanner(true);
-  };
-
-  if (!showAiGeneratedInputsBanner) {
-    return null;
-  }
-
-  return (
-    <div className="bg-primary/20 text-blue-300 text-sm px-2.5 py-2 rounded-md grid grid-cols-[auto_1fr_auto] gap-2 mb-4">
-      <div className="py-0.5">
-        <InfoCircledIcon className="w-3.5 h-3.5" />
-      </div>
-      <div className="flex flex-col items-start justify-start">
-        <span className="font-normal">Inputs generated by AI</span>
-        <span className="font-light"> Hit send to view output.</span>
-        <button
-          className="underline mt-2 font-light"
-          onClick={onDontShowAgain}
-          type="button"
-        >
-          Don&rsquo;t show again
-        </button>
-      </div>
-      <div>
-        <Button
-          className="p-0 h-auto"
-          variant="ghost"
-          size="sm"
-          onClick={onClose}
-        >
-          <Cross2Icon className="w-3.5 h-3.5 text-gray-400" />
-        </Button>
-      </div>
-    </div>
-  );
-}
-
 function isBodyEmpty(body: RequestorBody) {
   switch (body.type) {
     case "text":
@@ -485,164 +406,4 @@ function WebSocketNotConnectedBanner() {
       </div>
     </div>
   );
-}
-
-type AiDropDownMenuProps = {
-  isLoadingParameters: boolean;
-  persona: string;
-  onPersonaChange: (persona: AiTestingPersona) => void;
-  fillInRequest: () => void;
-};
-
-function AiDropDownMenu({
-  isLoadingParameters,
-  persona,
-  onPersonaChange,
-  fillInRequest,
-}: AiDropDownMenuProps) {
-  const [open, setOpen] = useState(false);
-
-  const handleValueChange = useCallback(
-    (value: string) => {
-      onPersonaChange(value === HOSTILE ? HOSTILE : FRIENDLY);
-    },
-    [onPersonaChange],
-  );
-
-  const handleGenerateRequest = useCallback(() => {
-    fillInRequest();
-    setOpen(false);
-  }, [fillInRequest, setOpen]);
-
-  // When the user shift+clicks of meta+clicks on the trigger,
-  // automatically open the menu
-  // I'm doing this because the caret is kinda hard to press...
-  const { isMetaOrShiftPressed } = useIsMetaOrShiftPressed();
-  const handleMagicWandButtonClick = useCallback(() => {
-    if (!open && isMetaOrShiftPressed) {
-      setOpen(true);
-      return;
-    }
-    fillInRequest();
-  }, [isMetaOrShiftPressed, setOpen, open, fillInRequest]);
-
-  return (
-    <DropdownMenu open={open} onOpenChange={setOpen}>
-      <Tooltip>
-        <TooltipTrigger asChild>
-          <div className="flex items-center">
-            <Button
-              variant="ghost"
-              className="p-2 h-auto"
-              size="sm"
-              onClick={handleMagicWandButtonClick}
-            >
-              <SparkleWand
-                className={cn("w-4 h-4", { "fpx-pulse": isLoadingParameters })}
-              />
-            </Button>
-            <DropdownMenuTrigger asChild>
-              <button>
-                <CaretDownIcon className="w-4 h-4" />
-              </button>
-            </DropdownMenuTrigger>
-          </div>
-        </TooltipTrigger>
-        <TooltipContent
-          className="bg-slate-900 px-2 py-1.5 text-white flex gap-1.5"
-          align="start"
-        >
-          Generate
-          <div className="flex gap-0.5">
-            <KeyboardShortcutKey>{isMac ? "⌘" : "Ctrl"}</KeyboardShortcutKey>{" "}
-            <KeyboardShortcutKey>G</KeyboardShortcutKey>
-          </div>
-        </TooltipContent>
-      </Tooltip>
-
-      <DropdownMenuContent className="min-w-60">
-        <DropdownMenuLabel>Generate Inputs</DropdownMenuLabel>
-        <DropdownMenuSeparator />
-        <DropdownMenuLabel className="font-normal">
-          Testing Persona
-        </DropdownMenuLabel>
-        <DropdownMenuRadioGroup
-          value={persona}
-          onValueChange={handleValueChange}
-        >
-          <DropdownMenuRadioItem
-            value="Friendly"
-            onClick={(e) => {
-              e.preventDefault();
-              e.stopPropagation();
-              onPersonaChange(FRIENDLY);
-            }}
-          >
-            Friendly
-          </DropdownMenuRadioItem>
-          <DropdownMenuRadioItem
-            value="QA"
-            onClick={(e) => {
-              e.preventDefault();
-              e.stopPropagation();
-              onPersonaChange(HOSTILE);
-            }}
-          >
-            Hostile
-          </DropdownMenuRadioItem>
-        </DropdownMenuRadioGroup>
-        <DropdownMenuSeparator />
-        <div className="px-2 py-1">
-          <Button
-            style={{
-              background: "linear-gradient(90deg, #3B82F6 0%, #C53BF6 100%)",
-            }}
-            className="w-full text-white flex gap-2 items-center"
-            // FIXME - While it's loading... show a spinner? And implement a timeout / cancel
-            disabled={isLoadingParameters}
-            onClick={handleGenerateRequest}
-          >
-            <SparkleWand className="w-4 h-4" />
-            <span>{isLoadingParameters ? "Generating..." : "Generate"}</span>
-          </Button>
-        </div>
-      </DropdownMenuContent>
-    </DropdownMenu>
-  );
-}
-
-function useIsMetaOrShiftPressed() {
-  const [isMetaOrShiftPressed, setIsMetaOrShiftPressed] = useState(false);
-
-  const handleKeyDown = useCallback((e: KeyboardEvent) => {
-    if (e.metaKey || e.shiftKey) {
-      setIsMetaOrShiftPressed(true);
-    }
-  }, []);
-
-  const handleKeyUp = useCallback((e: KeyboardEvent) => {
-    if (!e.metaKey && !e.shiftKey) {
-      setIsMetaOrShiftPressed(false);
-    }
-  }, []);
-
-  const handleBlur = useCallback(() => {
-    setIsMetaOrShiftPressed(false);
-  }, []);
-
-  useEffect(() => {
-    window.addEventListener("keydown", handleKeyDown);
-    window.addEventListener("keyup", handleKeyUp);
-    window.addEventListener("blur", handleBlur);
-
-    return () => {
-      window.removeEventListener("keydown", handleKeyDown);
-      window.removeEventListener("keyup", handleKeyUp);
-      window.removeEventListener("blur", handleBlur);
-    };
-  }, [handleKeyDown, handleKeyUp, handleBlur]);
-
-  return {
-    isMetaOrShiftPressed,
-  };
 }
